fix: align attention mask/bias dtype under half precision
DAT/HAT/SwinIR window attention added the float32 attention mask and
relative position bias directly onto a half-precision attn tensor, which
type-promoted the result back to float32 and then failed at the attn @ v
matmul with "expected scalar type Half but found Float" (same under bf16).
Cast both to attn.dtype before adding.

+++ cccv/arch/sr/dat_arch.py
@@ -425,14 +425,14 @@ class Spatial_Attention(nn.Module):
                 self.H_sp * self.W_sp, self.H_sp * self.W_sp, -1
             )
             relative_position_bias = relative_position_bias.permute(2, 0, 1).contiguous()
-            attn = attn + relative_position_bias.unsqueeze(0)
+            attn = attn + relative_position_bias.unsqueeze(0).to(attn.dtype)
 
         N = attn.shape[3]
 
         # use mask for shift window
         if mask is not None:
             nW = mask.shape[0]
-            attn = attn.view(B, nW, self.num_heads, N, N) + mask.unsqueeze(1).unsqueeze(0)
+            attn = attn.view(B, nW, self.num_heads, N, N) + mask.unsqueeze(1).unsqueeze(0).to(attn.dtype)
             attn = attn.view(-1, self.num_heads, N, N)
 
         attn = nn.functional.softmax(attn, dim=-1, dtype=attn.dtype)
